requirements.txt
django==1.8.7
pillow==3
django-celery==3.1.17